Rename configStore.js to configureStore.js, update imports
