+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from 'react';
 import { Provider } from 'react-redux';
-import configureStore from './store/configStore'
+import configureStore from './store/configureStore'
 
 import { HashRouter as Router } from 'react-router-dom';
