+++ client/src/containers/specs/FiltersCard.spec.tsx
@@ -5,7 +5,7 @@ import configureMockStore from '@jedmao/redux-mock-store';
 import { render, cleanup, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 
-import configureStore from '../../store/configStore';
+import configureStore from '../../store/configureStore';
 
 import FiltersCard from '../Nav/FiltersCard';
 import { DeepPartial } from 'redux';
